Premission fix for mssql

+++ src/NadekoBot/Modules/Permissions/PermissionExtensions.cs
@@ -106,13 +106,13 @@ namespace NadekoBot.Modules.Permissions
             switch (perm.PrimaryTarget)
             {
                 case PrimaryPermissionType.User:
-                    com += guild?.GetUser(perm.PrimaryTargetId).ToString() ?? $"<@{perm.PrimaryTargetId}>";
+                    com += guild?.GetUser((ulong) perm.PrimaryTargetId).ToString() ?? $"<@{perm.PrimaryTargetId}>";
                     break;
                 case PrimaryPermissionType.Channel:
                     com += $"<#{perm.PrimaryTargetId}>";
                     break;
                 case PrimaryPermissionType.Role:
-                    com += guild?.GetRole(perm.PrimaryTargetId)?.ToString() ?? $"<@&{perm.PrimaryTargetId}>";
+                    com += guild?.GetRole((ulong) perm.PrimaryTargetId)?.ToString() ?? $"<@&{perm.PrimaryTargetId}>";
                     break;
                 case PrimaryPermissionType.Server:
                     break;
